extract modeselector
